requirements.txt
astroid==2.1.0
boto3==1.9.86
botocore==1.12.86
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
psycopg2==2.7.7
pylint==2.2.2
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.11.1
